Flask-merge/requirements.txt
fastai==2.0.10
fastcore==1.0.9
fastprogress==1.0.0
Flask==1.1.2
Flask-RESTful==0.3.8
Pillow==7.2.0
simplejson==3.13.2
gunicorn==19.0.0
flask-cors
ipython==7.16.1
ipython-genutils==0.2.0
textblob==0.17.1
langdetect==1.0.9
urllib3
joblib
scikit-learn
nltk
keras
opencv-python
